license-plate-detection/.ipynb_checkpoints/requirements-checkpoint.txt
numpy
opencv-python-headless
ffmpeg-python